+++ tests/checks/entropy/test_calculate_entropy.py
@@ -4,7 +4,7 @@ Testing entropy functionality
 
 import pathlib
 
-from test_git_operations import get_most_recent_commits
+from tests.test_git import get_most_recent_commits
 
 from almanack.checks.entropy.calculate_entropy import (
     calculate_aggregate_entropy,
@@ -19,7 +19,7 @@ def test_calculate_normalized_entropy(
     Test the calculate_normalized_entropy function.
     """
     for label, repo_path in repository_paths.items():
-        # Extract two most recent commits: source and target
+        # Extract two most recent commits: source and tget_most_recent_commitsarget
         source_commit, target_commit = get_most_recent_commits(repo_path)
 
         # Call calculate_normalized_entropy function
