+++ tests/checks/entropy/test_compute_data.py
